requirements.txt
# Existing dependencies (as per your request)
fastapi>=0.104.0
uvicorn>=0.24.0
motor>=3.3.1
google-api-python-client>=2.106.0
python-dotenv>=1.0.0
pydantic>=2.4.2
beanie>=1.23.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
aiohttp==3.10
scikit-learn>=1.6.1
numpy>=2.2.4
pandas>=2.2.3
matplotlib>=3.10.1
scipy>=1.12.0
opencv-python>=4.9.0.80
playwright>=1.43.0
Pillow>=10.2.0
playwright-stealth>=1.0.6
setuptools>=69.2.0
python-dateutil>=2.8.2
